lavu/hwcontext_qsv: enable gpu_copy by default for hwdownload
Enable gpu_copy option for hwdownload by default, and can work with
single plane format(rgb32).

-For multiple plane format (like nv12),continuous memory allocated is needed;
-For single plane format(like rgb32), the default memory allocate method in FFmpeg is valid;

Default method in FFmpeg is to allocate continuous memory per plane, so the system memory allocated
for rgb32 is continuous, and can access the CM_mem_copy in MSDK if we set the option on.

Signed-off-by: Linjie Fu <linjie.fu@intel.com>